// dashboard/src/utils/hasura-api/generated/schemas/cronTriggerInvocationLogEntry.ts
/**
 * Generated by orval v8.0.3 🍺
 * Do not edit manually.
 * Hasura API
 * API for managing remote schemas and events in Hasura
 * OpenAPI spec version: 1.0.0
 */
import type { CronTriggerInvocationLogEntryRequest } from './cronTriggerInvocationLogEntryRequest';
import type { CronTriggerInvocationLogEntryResponse } from './cronTriggerInvocationLogEntryResponse';

export interface CronTriggerInvocationLogEntry {
  /** Represents a UUID for invocation. */
  id: string;
  /** Unique identifier for the event */
  event_id: string;
  /** Represents a timestamp value for the invocation of the trigger. */
  created_at: string;
  /**
   * HTTP status code of the webhook response
   * @nullable
   */
  status: number | null;
  /** Details of the HTTP request sent to the webhook */
  request: CronTriggerInvocationLogEntryRequest;
  /** Details of the HTTP response from the webhook */
  response?: CronTriggerInvocationLogEntryResponse;
}