// dashboard/src/utils/hasura-api/generated/schemas/getScheduledEventInvocationsArgs.ts
/**
 * Generated by orval v8.1.0 🍺
 * Do not edit manually.
 * Hasura API
 * API for managing remote schemas and events in Hasura
 * OpenAPI spec version: 1.0.0
 */
import type { GetScheduledEventInvocationsArgsType } from './getScheduledEventInvocationsArgsType';

export interface GetScheduledEventInvocationsArgs {
  /** Type of the Scheduled Trigger. */
  type: GetScheduledEventInvocationsArgsType;
  /** Name of the cron trigger. Required only when type is cron. */
  trigger_name?: string;
  /** Getting invocations of the specified event. */
  event_id?: string;
  /** Maximum number of events to be returned in one API call. */
  limit?: number;
  /** Starting point from where the rows need to be returned. */
  offset?: number;
  /** Flag to indicate whether number of rows returned should be included in the response. */
  get_rows_count?: boolean;
}